requirements.txt
beautifulsoup4>=4.12.3
openai>=1.93.0
python-dotenv>=1.0.1
requests>=2.32.3
